change sample code to use voltage detection;  bdb_commissioningNetworkFormation() issue for NON-FN; allow the ZED device enter ultra-low power mode while ble isn't busy for 5s

+++ telink_zigbee_ble_concurrent_sdk/apps/sampleSwitch/app_ui.c
@@ -153,7 +153,7 @@ void buttonKeepPressed(u8 btNum){
 	}
 }
 
-void set_detect_voltage(s32 v){
+void set_detect_voltage(u16 v){
     g_switchAppCtx.Vbat = v;
 }
